Created new base class that is inherited by all controls. This is so that the generic class can get the resource and then specific resources can interrogate it to provide a natural language test
Moved classes into aure_backend
